+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/discovery/PolarisDiscoveryHandler.java
@@ -28,28 +28,28 @@ import com.tencent.polaris.api.rpc.InstancesResponse;
 import com.tencent.polaris.api.rpc.ServicesResponse;
 import com.tencent.polaris.client.api.SDKContext;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
 /**
  * Discovery Handler for Polaris.
  *
  * @author Haotian Zhang, Andrew Shan, Jie Cheng
  */
-@Component
 public class PolarisDiscoveryHandler {
 
-	@Autowired
-	private PolarisDiscoveryProperties polarisDiscoveryProperties;
+	private final PolarisDiscoveryProperties polarisDiscoveryProperties;
 
-	@Autowired
-	private ProviderAPI providerAPI;
+	private final ProviderAPI providerAPI;
 
-	@Autowired
-	private SDKContext sdkContext;
+	private final SDKContext sdkContext;
 
-	@Autowired
-	private ConsumerAPI polarisConsumer;
+	private final ConsumerAPI polarisConsumer;
+
+	public PolarisDiscoveryHandler(PolarisDiscoveryProperties polarisDiscoveryProperties,
+				ProviderAPI providerAPI, SDKContext sdkContext, ConsumerAPI polarisConsumer) {
+		this.polarisDiscoveryProperties = polarisDiscoveryProperties;
+		this.providerAPI = providerAPI;
+		this.sdkContext = sdkContext;
+		this.polarisConsumer = polarisConsumer;
+	}
 
 	/**
 	 * Get a list of healthy instances.