+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/feign/plugin/reporter/SuccessPolarisReporter.java
@@ -30,7 +30,6 @@ import feign.Response;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.Ordered;
 import org.springframework.http.HttpStatus;
 
@@ -42,11 +41,12 @@ import org.springframework.http.HttpStatus;
 public class SuccessPolarisReporter extends AbstractPolarisReporterAdapter implements EnhancedFeignPlugin {
 
 	private static final Logger LOG = LoggerFactory.getLogger(SuccessPolarisReporter.class);
-	@Autowired(required = false)
-	private ConsumerAPI consumerAPI;
 
-	public SuccessPolarisReporter(RpcEnhancementReporterProperties properties) {
+	private final ConsumerAPI consumerAPI;
+
+	public SuccessPolarisReporter(RpcEnhancementReporterProperties properties, ConsumerAPI consumerAPI) {
 		super(properties);
+		this.consumerAPI = consumerAPI;
 	}
 
 	@Override