+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ---
 
 - [Optimize:optimize SpringValueProcessor.](https://github.com/Tencent/spring-cloud-tencent/pull/668)
+- [Optimize:optimize spring annotation.](https://github.com/Tencent/spring-cloud-tencent/pull/688)